code/requirements.txt
# Core dependencies 
pandas==2.3.3
numpy==2.3.5
tqdm==4.67.1

# LangChain + LLM connectors
langchain==1.1.0
langchain-community==0.4.1
langchain-openai==1.1.0
langchain-google-genai==3.2.0
langchain-core==1.1.0

# Vector store + embeddings
faiss-cpu==1.13.0

# PDF processing
PyPDF2==3.0.1

# Visualization and analysis
matplotlib==3.10.7
seaborn==0.13.2
altair==5.4.1
vl-convert-python==1.3.1
scikit-learn==1.7.2

# Environment management
python-dotenv==1.2.1

# Additional utilities
openpyxl==3.1.5

